TI-42 - feature(notify projects) : events return type

+++ smartimmo/src/main/java/com/gangdestrois/smartimmo/infrastructure/jpa/repository/NotificationRepository.java
@@ -2,10 +2,11 @@ package com.gangdestrois.smartimmo.infrastructure.jpa.repository;
 
 import com.gangdestrois.smartimmo.infrastructure.jpa.entity.NotificationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.validation.annotation.Validated;
 
 import java.util.List;
 
-public interface NotificationRepository extends JpaRepository<NotificationEntity, Long> {
+public interface NotificationRepository extends JpaRepository<NotificationEntity, Integer> {
     @Override
     List<NotificationEntity> findAll();
 }
